refactor: reorganize environment variable imports and add validation schemas for public and server environments
