+++ core/src/zap-old/hooks/push-notifications/use-push-notifications.ts
@@ -1,7 +1,7 @@
 "use client";
 import "client-only";
 
-import { PUBLIC_ENV } from "@/zap/env/env.public";
+import { PUBLIC_ENV } from "@/zap/env/public";
 import { handleClientError } from "@/zap/lib/api/client";
 import { ClientError, EnvironmentError } from "@/zap/lib/api/errors";
 import { useZapMutation } from "@/zap/lib/api/hooks/use-zap-mutation";
