+++ core/src/zap-old/lib/crypto/index.ts
@@ -1,6 +1,6 @@
 import "server-only";
 
-import { SERVER_ENV } from "@/zap/env/env.server";
+import { SERVER_ENV } from "@/zap/env/server";
 import { BadRequestError } from "@/zap/lib/api/errors";
 
 export const algorithm = "AES-CBC";
