+++ core/src/zap-old/lib/flags/flags.ts
@@ -1,6 +1,6 @@
 import { createPostHogAdapter } from "@flags-sdk/posthog";
 
-import { PUBLIC_ENV } from "@/zap/env/env.public";
+import { PUBLIC_ENV } from "@/zap/env/public";
 
 export const postHogAdapter = createPostHogAdapter({
   postHogKey: PUBLIC_ENV.NEXT_PUBLIC_POSTHOG_KEY || "",
